requirements.txt
Django==1.10.4
Pillow==3.4.2
django-classy-tags==0.7.2
django-cleanup==0.4.2
django-debug-toolbar==1.6
django-registration==2.2
django-registration-redux==1.4
elizabeth==0.3.17
psycopg2==2.6.2
sqlparse==0.2.2
